Rename PromotedSale component
