+++ desktop/apps/auction/components/Layout.js
@@ -1,5 +1,5 @@
 import ArtworksByFollowedArtists from 'desktop/apps/auction/components/artwork_browser/ArtworksByFollowedArtists'
-import BuyNowSaleArtworks from 'desktop/apps/auction/components/artwork_browser/BuyNowSaleArtworks'
+import PromotedSaleArtworks from 'desktop/apps/auction/components/artwork_browser/PromotedSaleArtworks'
 import ArtworkBrowser from 'desktop/apps/auction/components/artwork_browser/ArtworkBrowser'
 import AuctionBlock from 'desktop/components/react/auction_block/auction_block'
 import AuctionInfoContainer from 'desktop/apps/auction/components/layout/auction_info'
@@ -41,7 +41,7 @@ function Layout (props) {
           <MyActiveBids />
         }
 
-        <BuyNowSaleArtworks />
+        <PromotedSaleArtworks />
         <ArtworksByFollowedArtists />
 
         { showFilter && !showInfoWindow &&
