+++ desktop/apps/auction/components/artwork_browser/PromotedSaleArtworks.js
